[Refactor] Move KnobGestureRecord to KnobMapping.swift
