+++ Sources/Rings/KnobComponents/Mappings/KnobMapping.swift
@@ -5,7 +5,7 @@
 //  Created by Chen-Hai Teng on 2021/5/27.
 //
 
-import Foundation
+import SwiftUI
 
 enum Default {
     enum Degree: Double {
@@ -16,6 +16,16 @@ enum Default {
         case Min = 0.0
         case Max = 1.0
     }
+}
+
+public struct KnobGestureRecord {
+    struct Value {
+        var value: Double = .nan
+        var angle: Angle
+    }
+    var start: Value
+    var current: Value
+    var next: Value
 }
 
 public protocol KnobMapping : Adjustable {
